requirements.txt
Django
django-rest-framework
psycopg2-binary
Pillow
python-decouple
drf-yasg
